refactor(server): move RepositoryModule into models directory
